pyproject.toml
[tool.poetry]
name = "tellingme-python-server"
version = "0.1.0"
description = ""
authors = ["Dae-Jeong <v4chelsea@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
gunicorn = "^23.0.0"
greenlet = "^3.1.1"
python-multipart = "^0.0.12"
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.1"
cryptography = "^43.0.1"
httpx = "^0.27.2"
tortoise-orm = {version = "^0.21.6", extras = ["aiomysql"]}
celery = {extras = ["sqlalchemy"], version = "^5.4.0"}
apscheduler = "^3.11.0"
redis = "^5.2.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
mypy = "^1.11.2"
black = "^24.8.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
mypy-extensions = "^1.0.0"
coverage = "^7.6.8"
celery-stubs = "^0.1.3"

[tool.mypy]
files = "src"
strict = true


#disallow_untyped_calls = true  # 타입이 없는 함수 호출 금지
#disallow_untyped_defs = true   # 타입이 없는 함수 정의 금지
#ignore_missing_imports = true # 누락된 import 무시

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120


[tool.isort]
profile = "black"
line_length = 120

